refactor: split oversized modules by responsibility
