+++ crates/loonfs-cli/src/lib.rs
@@ -6,6 +6,7 @@
 
 mod args;
 mod backend;
+mod backend_error;
 mod commands;
 mod config;
 mod error;
