+++ crates/loonfs-core/src/path/write/planner.rs
@@ -1,22 +1,21 @@
-//! Plans one path mutation intent against namespace state, producing the
-//! commit plan and semantic fingerprint the publish path needs.
+//! Path-intent fingerprinting and dispatch to responsibility-specific planners.
 
 use super::intent::PathMutationIntent;
+use super::plan_create::{
+    plan_publish_create_directory, plan_publish_put_file_content_ref, plan_publish_undelete,
+};
+use super::plan_delete::plan_publish_delete_path;
+use super::plan_restore::plan_publish_restore_revision;
+use super::plan_transfer::{plan_publish_copy_file_path, plan_publish_move_path};
+use super::planning_helpers::PublishPathPlanningView;
 use crate::commit::{fingerprint_digest, PathIntentFingerprint, PATH_INTENT_FINGERPRINT_DOMAIN};
 use crate::error::{CoreError, Result};
-#[cfg(test)]
-use crate::metadata::MetadataState;
-use crate::metadata::{MetadataView, ResolvedVisiblePath, VisiblePathError};
-use crate::path::helpers::{ensure_mutation_path, final_component};
+use crate::metadata::MetadataView;
 use loonfs_api::wire::control::HeadState;
 use loonfs_api::ChangeSeq;
 use loonfs_api::{
-    v0::{
-        CommitOp as ApiCommitOp, CommitPrecondition as ApiCommitPrecondition,
-        CommitRequest as ApiCommitRequest,
-    },
-    AbsolutePath, CommitId, ContentRef, DeleteDirectoryBehavior, DestinationBehavior, DisplayName,
-    InodeId, InodeKind, NameKey, NamespaceId, RevisionNo, ROOT_INODE_ID,
+    v0::CommitRequest as ApiCommitRequest, CommitId, ContentRef, DeleteDirectoryBehavior,
+    DestinationBehavior, InodeId, NamespaceId, RevisionNo,
 };
 use loonfs_objectstore::ObjectStore;
 use serde::Serialize;
@@ -181,13 +180,6 @@ pub(crate) fn path_intent_fingerprint(
     fingerprint_path_input(&identity)
 }
 
-fn is_missing_visible_path(error: &CoreError) -> bool {
-    matches!(
-        error,
-        CoreError::PathNotFound(_) | CoreError::VisiblePath(VisiblePathError::PathNotFound { .. })
-    )
-}
-
 pub(crate) async fn plan_path_mutation_against_publish_view<S: ObjectStore + ?Sized>(
     namespace_id: &NamespaceId,
     intent: &PathMutationIntent,
@@ -273,663 +265,18 @@ pub(crate) async fn plan_path_mutation_against_publish_view<S: ObjectStore + ?Si
     })
 }
 
-struct PublishPathPlanningView<'a, 'b, 'store, S: ObjectStore + ?Sized> {
-    head: &'a HeadState,
-    metadata_state: &'a MetadataView<'b, 'store, S>,
-}
-
-async fn publish_binding_is_precondition<S: ObjectStore + ?Sized>(
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-    resolved: &ResolvedVisiblePath,
-) -> Result<ApiCommitPrecondition> {
-    let parent_inode_id = resolved
-        .parent_inode_id
-        .ok_or(CoreError::RootMutationForbidden)?;
-    let binding = view
-        .metadata_state
-        .current_parent_binding_for_child(resolved.inode_id)
-        .await?
-        .ok_or_else(|| CoreError::PathNotFound(resolved.absolute_path.clone()))?;
-    if binding.parent_inode_id != parent_inode_id {
-        return Err(CoreError::PathNotFound(resolved.absolute_path.clone()));
-    }
-    Ok(ApiCommitPrecondition::BindingIs {
-        parent_inode_id,
-        name_key: binding.name_key.clone(),
-        child_inode_id: binding.child_inode_id,
-        bind_seq: binding.bind_seq,
-        bind_delta_index: binding.bind_delta_index,
-    })
-}
-
-fn publish_child_name_absent_precondition<S: ObjectStore + ?Sized>(
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-    parent_inode_id: InodeId,
-    display_name: &DisplayName,
-) -> ApiCommitPrecondition {
-    let name_key = NameKey::for_display_name(view.metadata_state.name_policy(), display_name);
-    ApiCommitPrecondition::ChildNameAbsent {
-        parent_inode_id,
-        name_key,
-    }
-}
-
-/// Rejects planning through a *visible* path component covered by a subtree
-/// tombstone. The walk observes only visible bindings, so its answer cannot
-/// change when compaction drops rows no retained sequence observes: a deleted
-/// (unbound) name simply ends the walk, and recreating it plans as a fresh
-/// subtree. A visible-but-covered component cannot arise from legal writer
-/// histories; hitting one means corrupt state, and failing the plan is the
-/// safe answer.
-async fn publish_reject_tombstoned_path_ancestor<S: ObjectStore + ?Sized>(
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-    absolute_path: &AbsolutePath,
-) -> Result<()> {
-    let mut current_inode = ROOT_INODE_ID;
-    let mut current_path = AbsolutePath::root();
-
-    for component in absolute_path.components() {
-        let display_name = component.to_display_name();
-        let name_key = NameKey::for_display_name(view.metadata_state.name_policy(), &display_name);
-        let Some(bound_child) = view
-            .metadata_state
-            .visible_child(current_inode, &name_key)
-            .await?
-        else {
-            return Ok(());
-        };
-        let visible_component = bound_child.display_name.clone();
-        let visible_path = current_path.join(&visible_component);
-        if let Some(tombstone) = view
-            .metadata_state
-            .covering_subtree_tombstone(bound_child.child_inode_id)
-            .await?
-        {
-            return Err(CoreError::TombstoneConflict {
-                path: visible_path.as_str().to_owned(),
-                root_inode_id: tombstone.root_inode_id,
-                tombstone_seq: tombstone.tombstone_seq,
-            });
-        }
-        current_inode = bound_child.child_inode_id;
-        current_path = visible_path;
-    }
-    Ok(())
-}
-
-async fn plan_publish_create_directory<S: ObjectStore + ?Sized>(
-    absolute_path: &AbsolutePath,
-    parents: bool,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(absolute_path)?;
-    publish_reject_tombstoned_path_ancestor(view, absolute_path).await?;
-    match view
-        .metadata_state
-        .resolve_visible_path(absolute_path)
-        .await
-    {
-        Ok(_) => {
-            return Err(CoreError::DestinationExists(
-                absolute_path.as_str().to_owned(),
-            ));
-        }
-        Err(error) if is_missing_visible_path(&error) => {}
-        Err(error) => return Err(error),
-    }
-    let mut ops = Vec::new();
-    let parent_inode_id = if parents {
-        // The same ancestor auto-create the put-file plan performs.
-        let mut next_inode_id = view.head.next_inode_id;
-        publish_ensure_parent_directories(absolute_path, view, &mut ops, &mut next_inode_id).await?
-    } else {
-        publish_resolve_parent_directory(view, absolute_path).await?
-    };
-    let display_name = final_component(absolute_path)?;
-    ops.push(ApiCommitOp::CreateDirectory {
-        parent_inode_id,
-        display_name: display_name.clone(),
-    });
-    // A parent allocated by this same commit cannot have conflicting
-    // children yet, so the name and ancestor preconditions only apply when
-    // the parent already exists — mirroring the put-file plan.
-    let mut preconditions = Vec::new();
-    if view
-        .metadata_state
-        .visible_inode(parent_inode_id)
-        .await?
-        .is_some()
-    {
-        preconditions.push(publish_child_name_absent_precondition(
-            view,
-            parent_inode_id,
-            &display_name,
-        ));
-        preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-            inode_id: parent_inode_id,
-        });
-    }
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops,
-        preconditions,
-        message: None,
-    })
-}
-
-async fn plan_publish_undelete<S: ObjectStore + ?Sized>(
-    inode_id: InodeId,
-    deleted_at_seq: ChangeSeq,
-    absolute_path: &AbsolutePath,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(absolute_path)?;
-    publish_reject_tombstoned_path_ancestor(view, absolute_path).await?;
-    match view
-        .metadata_state
-        .resolve_visible_path(absolute_path)
-        .await
-    {
-        Ok(_) => {
-            return Err(CoreError::DestinationExists(
-                absolute_path.as_str().to_owned(),
-            ));
-        }
-        Err(error) if is_missing_visible_path(&error) => {}
-        Err(error) => return Err(error),
-    }
-    // The destination parent must already exist: recovery targets a place
-    // the caller can see, and commit validation re-checks the tombstone
-    // root, the parent, and the name under the publish lock.
-    let parent_inode_id = publish_resolve_parent_directory(view, absolute_path).await?;
-    let display_name = final_component(absolute_path)?;
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops: vec![ApiCommitOp::Undelete {
-            inode_id,
-            deleted_at_seq,
-            parent_inode_id,
-            display_name: display_name.clone(),
-        }],
-        preconditions: vec![
-            publish_child_name_absent_precondition(view, parent_inode_id, &display_name),
-            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                inode_id: parent_inode_id,
-            },
-        ],
-        message: None,
-    })
-}
-
-async fn plan_publish_put_file_content_ref<S: ObjectStore + ?Sized>(
-    absolute_path: &AbsolutePath,
-    content_ref: ContentRef,
-    behavior: DestinationBehavior,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(absolute_path)?;
-    publish_reject_tombstoned_path_ancestor(view, absolute_path).await?;
-    let target = view
-        .metadata_state
-        .resolve_visible_path(absolute_path)
-        .await;
-
-    let mut ops = Vec::new();
-    let mut next_inode_id = view.head.next_inode_id;
-    let final_parent_inode =
-        publish_ensure_parent_directories(absolute_path, view, &mut ops, &mut next_inode_id)
-            .await?;
-    let final_name = final_component(absolute_path)?;
-    let mut preconditions = Vec::new();
-
-    match target {
-        Ok(existing) => {
-            if behavior == DestinationBehavior::NoReplace {
-                return Err(CoreError::DestinationExists(
-                    absolute_path.as_str().to_owned(),
-                ));
-            }
-            if existing.inode_kind != InodeKind::File {
-                return Err(CoreError::ExpectedFile {
-                    path: absolute_path.as_str().to_owned(),
-                    kind: existing.inode_kind,
-                });
-            }
-            let revision = view
-                .metadata_state
-                .latest_revision_head(existing.inode_id)
-                .await?
-                .ok_or_else(|| CoreError::PathNotFound(absolute_path.as_str().to_owned()))?;
-            preconditions.push(publish_binding_is_precondition(view, &existing).await?);
-            ops.push(ApiCommitOp::ReplaceFile {
-                inode_id: existing.inode_id,
-                base_revision_no: revision.revision_no,
-                content_ref: content_ref.clone(),
-            });
-            preconditions.push(ApiCommitPrecondition::InodeRevisionIs {
-                inode_id: existing.inode_id,
-                revision_no: revision.revision_no,
-            });
-            preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                inode_id: existing.inode_id,
-            });
-        }
-        Err(error) if is_missing_visible_path(&error) => {
-            ops.push(ApiCommitOp::CreateFile {
-                parent_inode_id: final_parent_inode,
-                display_name: final_name.clone(),
-                content_ref,
-            });
-            if view
-                .metadata_state
-                .visible_inode(final_parent_inode)
-                .await?
-                .is_some()
-            {
-                preconditions.push(publish_child_name_absent_precondition(
-                    view,
-                    final_parent_inode,
-                    &final_name,
-                ));
-                preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                    inode_id: final_parent_inode,
-                });
-            }
-        }
-        Err(other) => return Err(other),
-    }
-
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops,
-        preconditions,
-        message: None,
-    })
-}
-
-async fn plan_publish_delete_path<S: ObjectStore + ?Sized>(
-    absolute_path: &AbsolutePath,
-    behavior: DeleteDirectoryBehavior,
-    expected_inode_id: Option<InodeId>,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(absolute_path)?;
-    let resolved = view
-        .metadata_state
-        .resolve_visible_path(absolute_path)
-        .await?;
-    // Planning happens under the publish lock, so this check is race-free:
-    // a caller that resolved the path earlier (a stat) either deletes that
-    // exact inode or fails, never a raced rebinding.
-    if let Some(expected) = expected_inode_id {
-        if resolved.inode_id != expected {
-            return Err(
-                crate::commit::CommitValidationError::BindingPreconditionMismatch {
-                    // Root cannot be deleted, so a resolved delete target
-                    // always has a parent.
-                    parent_inode_id: resolved.parent_inode_id.unwrap_or(ROOT_INODE_ID),
-                    name_key: NameKey::for_display_name(
-                        view.metadata_state.name_policy(),
-                        &absolute_path
-                            .final_component()
-                            .expect("non-root mutation path should have a final component")
-                            .to_display_name(),
-                    ),
-                    expected_child_inode_id: expected,
-                    actual_child_inode_id: resolved.inode_id,
-                }
-                .into(),
-            );
-        }
-    }
-    let recursive = behavior == DeleteDirectoryBehavior::Recursive;
-    let op = match resolved.inode_kind {
-        InodeKind::File => ApiCommitOp::DeleteFile {
-            inode_id: resolved.inode_id,
-        },
-        InodeKind::Directory if recursive => ApiCommitOp::DeleteSubtree {
-            root_inode_id: resolved.inode_id,
-        },
-        InodeKind::Directory => {
-            if view
-                .metadata_state
-                .has_visible_children(resolved.inode_id)
-                .await?
-            {
-                return Err(CoreError::DirectoryNotEmpty(
-                    absolute_path.as_str().to_owned(),
-                ));
-            }
-            ApiCommitOp::DeleteSubtree {
-                root_inode_id: resolved.inode_id,
-            }
-        }
-    };
-    let mut preconditions = vec![publish_binding_is_precondition(view, &resolved).await?];
-    if !recursive && resolved.inode_kind == InodeKind::Directory {
-        preconditions.push(ApiCommitPrecondition::DirectoryEmpty {
-            inode_id: resolved.inode_id,
-        });
-    }
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops: vec![op],
-        preconditions,
-        message: None,
-    })
-}
-
-async fn plan_publish_move_path<S: ObjectStore + ?Sized>(
-    from_path: &AbsolutePath,
-    to_path: &AbsolutePath,
-    behavior: DestinationBehavior,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(from_path)?;
-    ensure_mutation_path(to_path)?;
-    publish_reject_tombstoned_path_ancestor(view, from_path).await?;
-    publish_reject_tombstoned_path_ancestor(view, to_path).await?;
-    let source = view.metadata_state.resolve_visible_path(from_path).await?;
-    let target_parent = publish_resolve_parent_directory(view, to_path).await?;
-    let target_name = final_component(to_path)?;
-    // Replace compiles to an atomic delete-plus-rename: the destination
-    // file's delete and the source's rebind land in one commit, and the
-    // rename's target-name check observes the in-commit unbind. Mirrors
-    // put: only a file destination can be replaced, and a path never
-    // replaces itself.
-    let replaced = match view.metadata_state.resolve_visible_path(to_path).await {
-        Ok(existing)
-            if behavior == DestinationBehavior::Replace && existing.inode_id != source.inode_id =>
-        {
-            if existing.inode_kind != InodeKind::File {
-                return Err(CoreError::ExpectedFile {
-                    path: to_path.as_str().to_owned(),
-                    kind: existing.inode_kind,
-                });
-            }
-            Some(existing)
-        }
-        Ok(_) => return Err(CoreError::DestinationExists(to_path.as_str().to_owned())),
-        Err(error) if is_missing_visible_path(&error) => None,
-        Err(error) => return Err(error),
-    };
-    let mut ops = Vec::new();
-    let mut preconditions = vec![publish_binding_is_precondition(view, &source).await?];
-    match &replaced {
-        Some(existing) => {
-            ops.push(ApiCommitOp::DeleteFile {
-                inode_id: existing.inode_id,
-            });
-            preconditions.push(publish_binding_is_precondition(view, existing).await?);
-            preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                inode_id: existing.inode_id,
-            });
-        }
-        None => {
-            preconditions.push(publish_child_name_absent_precondition(
-                view,
-                target_parent,
-                &target_name,
-            ));
-        }
-    }
-    ops.push(ApiCommitOp::Rename {
-        inode_id: source.inode_id,
-        new_parent_inode_id: target_parent,
-        new_display_name: target_name.clone(),
-    });
-    preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-        inode_id: source.inode_id,
-    });
-    preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-        inode_id: target_parent,
-    });
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops,
-        preconditions,
-        message: None,
-    })
-}
-
-async fn plan_publish_copy_file_path<S: ObjectStore + ?Sized>(
-    from_path: &AbsolutePath,
-    to_path: &AbsolutePath,
-    behavior: DestinationBehavior,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(from_path)?;
-    ensure_mutation_path(to_path)?;
-    publish_reject_tombstoned_path_ancestor(view, from_path).await?;
-    publish_reject_tombstoned_path_ancestor(view, to_path).await?;
-
-    let source = view.metadata_state.resolve_visible_path(from_path).await?;
-    if source.inode_kind != InodeKind::File {
-        return Err(CoreError::ExpectedFile {
-            path: from_path.as_str().to_owned(),
-            kind: source.inode_kind,
-        });
-    }
-
-    // Replace mirrors put onto an existing file: the copy appends a new
-    // revision to the destination inode, keeping its identity and revision
-    // history. Only a file destination can be replaced, and a path never
-    // replaces itself.
-    let replaced = match view.metadata_state.resolve_visible_path(to_path).await {
-        Ok(existing)
-            if behavior == DestinationBehavior::Replace && existing.inode_id != source.inode_id =>
-        {
-            if existing.inode_kind != InodeKind::File {
-                return Err(CoreError::ExpectedFile {
-                    path: to_path.as_str().to_owned(),
-                    kind: existing.inode_kind,
-                });
-            }
-            Some(existing)
-        }
-        Ok(_) => return Err(CoreError::DestinationExists(to_path.as_str().to_owned())),
-        Err(error) if is_missing_visible_path(&error) => None,
-        Err(error) => return Err(error),
-    };
-
-    let revision = view
-        .metadata_state
-        .latest_revision_head(source.inode_id)
-        .await?
-        .ok_or_else(|| CoreError::PathNotFound(from_path.as_str().to_owned()))?;
-
-    let target_parent = publish_resolve_parent_directory(view, to_path).await?;
-    let target_name = final_component(to_path)?;
-    let mut ops = Vec::new();
-    let mut preconditions = vec![
-        publish_binding_is_precondition(view, &source).await?,
-        ApiCommitPrecondition::InodeRevisionIs {
-            inode_id: source.inode_id,
-            revision_no: revision.revision_no,
-        },
-        ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-            inode_id: source.inode_id,
-        },
-    ];
-    match &replaced {
-        Some(existing) => {
-            let existing_revision = view
-                .metadata_state
-                .latest_revision_head(existing.inode_id)
-                .await?
-                .ok_or_else(|| CoreError::PathNotFound(to_path.as_str().to_owned()))?;
-            ops.push(ApiCommitOp::ReplaceFile {
-                inode_id: existing.inode_id,
-                base_revision_no: existing_revision.revision_no,
-                content_ref: revision.content_ref,
-            });
-            preconditions.push(publish_binding_is_precondition(view, existing).await?);
-            preconditions.push(ApiCommitPrecondition::InodeRevisionIs {
-                inode_id: existing.inode_id,
-                revision_no: existing_revision.revision_no,
-            });
-            preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                inode_id: existing.inode_id,
-            });
-        }
-        None => {
-            ops.push(ApiCommitOp::CreateFile {
-                parent_inode_id: target_parent,
-                display_name: target_name.clone(),
-                content_ref: revision.content_ref,
-            });
-            preconditions.push(publish_child_name_absent_precondition(
-                view,
-                target_parent,
-                &target_name,
-            ));
-        }
-    }
-    preconditions.push(ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-        inode_id: target_parent,
-    });
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops,
-        preconditions,
-        message: None,
-    })
-}
-
-async fn plan_publish_restore_revision<S: ObjectStore + ?Sized>(
-    absolute_path: &AbsolutePath,
-    source_revision_no: RevisionNo,
-    commit_id: &CommitId,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-) -> Result<ApiCommitRequest> {
-    ensure_mutation_path(absolute_path)?;
-    publish_reject_tombstoned_path_ancestor(view, absolute_path).await?;
-    let target = view
-        .metadata_state
-        .resolve_visible_path(absolute_path)
-        .await?;
-    if target.inode_kind != InodeKind::File {
-        return Err(CoreError::ExpectedFile {
-            path: absolute_path.as_str().to_owned(),
-            kind: target.inode_kind,
-        });
-    }
-    let revision = view
-        .metadata_state
-        .latest_revision_head(target.inode_id)
-        .await?
-        .ok_or_else(|| CoreError::PathNotFound(absolute_path.as_str().to_owned()))?;
-
-    Ok(ApiCommitRequest {
-        commit_id: commit_id.to_owned(),
-        ops: vec![ApiCommitOp::RestoreRevision {
-            inode_id: target.inode_id,
-            source_revision_no,
-            base_revision_no: revision.revision_no,
-        }],
-        preconditions: vec![
-            publish_binding_is_precondition(view, &target).await?,
-            ApiCommitPrecondition::InodeRevisionIs {
-                inode_id: target.inode_id,
-                revision_no: revision.revision_no,
-            },
-            ApiCommitPrecondition::AncestorsNotSubtreeDeleted {
-                inode_id: target.inode_id,
-            },
-        ],
-        message: None,
-    })
-}
-
-async fn publish_ensure_parent_directories<S: ObjectStore + ?Sized>(
-    absolute_path: &AbsolutePath,
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-    ops: &mut Vec<ApiCommitOp>,
-    next_inode_id: &mut InodeId,
-) -> Result<InodeId> {
-    let components = absolute_path.components();
-    if components.len() <= 1 {
-        return Ok(ROOT_INODE_ID);
-    }
-
-    let mut current_inode = ROOT_INODE_ID;
-    let mut creating_missing_ancestors = false;
-    for component in &components[..components.len() - 1] {
-        let display_name = component.to_display_name();
-        let name_key = NameKey::for_display_name(view.metadata_state.name_policy(), &display_name);
-        if !creating_missing_ancestors {
-            if let Some(child) = view
-                .metadata_state
-                .visible_child(current_inode, &name_key)
-                .await?
-            {
-                let inode = view
-                    .metadata_state
-                    .visible_inode(child.child_inode_id)
-                    .await?
-                    .ok_or_else(|| CoreError::PathNotFound(component.as_str().to_owned()))?;
-                if inode.inode_kind != InodeKind::Directory {
-                    return Err(CoreError::NonDirectoryPathComponent(
-                        component.as_str().to_owned(),
-                    ));
-                }
-                current_inode = child.child_inode_id;
-                continue;
-            }
-            creating_missing_ancestors = true;
-        }
-
-        ops.push(ApiCommitOp::CreateDirectory {
-            parent_inode_id: current_inode,
-            display_name,
-        });
-        let allocated = *next_inode_id;
-        *next_inode_id = InodeId(next_inode_id.0.saturating_add(1));
-        current_inode = allocated;
-    }
-    Ok(current_inode)
-}
-
-async fn publish_resolve_parent_directory<S: ObjectStore + ?Sized>(
-    view: &PublishPathPlanningView<'_, '_, '_, S>,
-    absolute_path: &AbsolutePath,
-) -> Result<InodeId> {
-    let Some(parent_path) = absolute_path.parent() else {
-        return Ok(ROOT_INODE_ID);
-    };
-    if parent_path.is_root() {
-        return Ok(ROOT_INODE_ID);
-    }
-    let resolved = view
-        .metadata_state
-        .resolve_visible_path(&parent_path)
-        .await?;
-    if resolved.inode_kind != InodeKind::Directory {
-        return Err(CoreError::ExpectedDirectory {
-            path: parent_path.as_str().to_owned(),
-            kind: resolved.inode_kind,
-        });
-    }
-    Ok(resolved.inode_id)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::commit::core_commit_fingerprint_for_v0_request;
     use crate::context::MutationContext;
+    use crate::metadata::MetadataState;
     use crate::namespace::bootstrap::bootstrap_namespace;
     use crate::path::write::ops::{delete_path, put_file_bytes};
     use crate::protocol::{load_publish_metadata_view, PublishTailOptions};
     use crate::storage::content::store_bytes_as_content;
     use loonfs_api::v0::{CommitOp, CommitPrecondition, CommitRequest as ApiCommitRequest};
-    use loonfs_api::RevisionNo;
+    use loonfs_api::{AbsolutePath, RevisionNo};
     use loonfs_objectstore::local_fs_store::LocalFsStore;
     use tempfile::tempdir;
